requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
decorator==4.4.2
Django==3.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
moviepy==1.0.3
numpy==1.20.3
Pillow==8.0.1
proglog==0.1.9
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.5
waitress==1.4.4
whitenoise==5.2.0
